Refactored functions to use async thunks
